Added new dependency: redis
