+++ requirements.txt
@@ -10,3 +10,4 @@ pydantic-settings~=2.3.4
 pytest~=8.3.2
 python-dotenv~=1.0.1
 python-jose~=3.3.0
+redis~=6.2.0
